requirements.txt
Flask==3.0.2
numpy==1.24.2
pymongo==4.6.1
tensorflow-cpu==2.15.0
spacy==3.7.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl